pyproject.toml
[project]
name = "python-backup-script"
version = "0.1.0"
description = "This project have the objective of making periodically database backups on the production server, sending files to a backup server via SSH"
authors = [
    {name = "gabrielms-castro",email = "gabriel.castro@phddobrasil.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "paramiko (>=4.0.0,<5.0.0)",
    "scp (>=0.15.0,<0.16.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.ruff]
line-length = 79
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".gitignore",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic",
    "migrations",    
]


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR2004', 'PLR0917', 'PLR0913']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'python3 app/main.py'
test = 'python3 -m unittest discover -s tests'